+++ src/ml/GenerateSchemaTypes.ml
@@ -27,6 +27,50 @@ type fieldResolverStyle =
   | Resolver of {moduleName: string; fnName: string; pathToFn: string list}
   | Property of string
 
+type authorizationMode = AuthorizationOptional | AuthorizationRequired
+
+type authorizationConfig = {
+  mode: authorizationMode;
+  onForbidden: string option;
+  manifestPath: string option;
+}
+
+type authorizationFunctionReference = {
+  path: string list;
+  loc: Location.t;
+  fileUri: Uri.t;
+}
+
+type publicAuthorization = {reason: string; loc: Location.t; fileUri: Uri.t}
+
+type declaredAuthorization = {
+  functions: authorizationFunctionReference list;
+  public: publicAuthorization option;
+}
+
+type authorizationInjection = AuthorizationContext | AuthorizationInfo
+
+type authorizationProvenance =
+  | ObjectTypePolicy of string
+  | InterfaceTypePolicy of string
+  | InterfaceFieldPolicy of string
+  | FieldPolicy of string
+
+type authorizationFunction = {
+  reference: authorizationFunctionReference;
+  isAsync: bool;
+  injections: authorizationInjection list;
+  provenance: authorizationProvenance;
+}
+
+type resolverOutcome = {isAsync: bool}
+
+type effectiveAuthorizationPlan = {
+  functions: authorizationFunction list;
+  public: publicAuthorization option;
+  resolverOutcome: resolverOutcome option;
+}
+
 type typeLocationLoc = {
   fileName: string;
   fileUri: Uri.t;
@@ -172,6 +216,10 @@ type schemaState = {
   interfaces: (string, gqlInterface) Hashtbl.t;
   scalars: (string, gqlScalar) Hashtbl.t;
   processedFiles: (string, bool) Hashtbl.t;
+  authorizationConfig: authorizationConfig;
+  authorizationDeclarations: (string, declaredAuthorization) Hashtbl.t;
+  authorizationPlans: (string, effectiveAuthorizationPlan) Hashtbl.t;
+  resolverOutcomes: (string, resolverOutcome) Hashtbl.t;
   mutable query: gqlObjectType option;
   mutable subscription: gqlObjectType option;
   mutable mutation: gqlObjectType option;